refactor(src/Utils): Move network constants from connectors to utils/constants
