+++ src/utils/constants.ts
@@ -6,3 +6,30 @@ export const ANY_FUNC_SIGNATURE = "0xaaaaaaaa";
 export const ERC20_TRANSFER_SIGNATURE = "0xa9059cbb";
 export const ERC20_APPROVE_SIGNATURE = "0x095ea7b3";
 export const DEFAULT_TOKEN_DECIMALS = 18;
+
+export const NETWORK_NAMES = {
+  '1': 'mainnet',
+  '4': 'rinkeby',
+  '100': 'xdai',
+  '42161': 'arbitrum',
+  '421611': 'arbitrumTestnet',
+  '1337': 'localhost',
+};
+
+export const NETWORK_IDS = {
+  'mainnet': 1,
+  'rinkeby': 4,
+  'xdai': 100,
+  'arbitrum': 42161,
+  'arbitrumTestnet': 421611,
+  'localhost': 1337,
+};
+
+export const NETWORK_ASSET_SYMBOL = {
+  'mainnet': 'ETH',
+  'rinkeby': 'ETH',
+  'xdai': 'XDAI',
+  'arbitrum': 'ETH',
+  'arbitrumTestnet': 'ETH',
+  'localhost': 'ETH'
+};
